+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ~~~~~~~~~
 
+0.32.0 (January 2024)
+---------------------
+
+* Updated pytket version requirement to 1.23.
+* Don’t include :py:class:`SimplifyInitial` in default passes; instead make it an option to ``process_circuits()``.
+
 0.31.0 (November 2023)
 ----------------------